chore: setup specs to support SDK targeting
to support an array of expo SDK versions we need to allow for running tests with an SDK major version input

+++ test/setup.sh
@@ -2,11 +2,17 @@
 
 set -e
 
+if [ -z "$EXPO_SDK_TARGET" ]; then
+  echo "Tests must specify EXPO_SDK_TARGET"
+  echo ""
+  exit 1
+fi
+
 npm run build
 
 echo "Creating expo app"
 rm -rf example
-CI=1 npx create-expo-app example
+CI=1 npx create-expo-app --template "expo-template-default@$EXPO_SDK_TARGET" example
 
 echo "Installing library"
 cd example
